fix for error message tkversion, required some restructuring of stypes/exceptions/libspicehelper
