+++ src/spiceypy/spiceypy.py
@@ -36,6 +36,14 @@ from .utils import support_types as stypes
 from .utils.libspicehelper import libspice
 from . import config
 
+from .utils.exceptions import *
+
+# inject exceptions back into stypes for backwards compatibility
+stypes.SpiceyError = SpiceyError
+stypes.SpiceyPyError = SpiceyPyError
+stypes.NotFoundError = NotFoundError
+stypes.exceptions = exceptions
+
 from .utils.callbacks import (
     UDFUNC,
     UDFUNS,
@@ -104,7 +112,7 @@ def check_for_spice_error(f: Optional[Callable]) -> None:
         long = getmsg("LONG", 1841).strip()
         traceback = qcktrc(200)
         reset()
-        raise stypes.dynamically_instantiate_spiceyerror(
+        raise dynamically_instantiate_spiceyerror(
             short=short, explain=explain, long=long, traceback=traceback
         )
 
@@ -140,12 +148,12 @@ def spice_found_exception_thrower(f: Callable) -> Callable:
         if config.catch_false_founds:
             found = res[-1]
             if isinstance(found, bool) and not found:
-                raise stypes.NotFoundError(
+                raise NotFoundError(
                     "Spice returns not found for function: {}".format(f.__name__),
                     found=found,
                 )
             elif stypes.is_iterable(found) and not all(found):
-                raise stypes.NotFoundError(
+                raise NotFoundError(
                     "Spice returns not found in a series of calls for function: {}".format(
                         f.__name__
                     ),
